pyproject.toml
[project]
name = "JobSage"
version = "1.0.0"
description = "Pakistan's first data-driven career intelligence platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "znpg>=1.1.1",
    "playwright>=1.30.0",
    'python-dotenv>=1.0.0',
    'pytest>=7.4.0',
    "fastapi[standard]>=0.128.0",
    "uvicorn>=0.40.0",
    "apscheduler>=3.11.2"
]
